[win32] Fix pointer to integer casts
'long' will be 32b on Windows, even on x64 targets. For this reason,
we can't use "long" when casting a pointer.

We'll use uintptr_t instead.

+++ src/include/buffer.h
@@ -292,7 +292,7 @@ inline namespace v14_2_0 {
 
     // misc
     bool is_aligned(unsigned align) const {
-      return ((long)c_str() & (align-1)) == 0;
+      return ((uintptr_t)c_str() & (align-1)) == 0;
     }
     bool is_page_aligned() const { return is_aligned(CEPH_PAGE_SIZE); }
     bool is_n_align_sized(unsigned align) const
